Added travis, fixed tests.
