+++ src/main/java/lastunion/application/controllers/SignUpController.java
@@ -14,7 +14,7 @@ import javax.servlet.http.HttpSession;
 import javax.validation.constraints.NotNull;
 import java.util.Locale;
 
-//@CrossOrigin(origins = "https://front-lastunion.herokuapp.com/")
+@CrossOrigin(origins = "https://front-lastunion.herokuapp.com/")
 @RestController
 public class SignUpController {
     @NotNull
