+++ src/main/resources/application.properties
@@ -3,5 +3,5 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
 security.basic.enabled=false
 # suppress inspection "SpringBootApplicationProperties" ${JDBC_DATABASE_URL}
 spring.datasource.initialSize=1
-logging.level.root=debug
+logging.level.root=error
 #server.port= 9000
