+++ src/test/java/tests/Test/SignUpTest_IT.java
@@ -23,9 +23,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest(classes = Application.class)
 @AutoConfigureMockMvc(print = MockMvcPrint.NONE)
 @Category(tests.IntegrationTest.class)
-//
-
-public class ChangePaswordTest {
+public class SignUpTest_IT {
     @Autowired
     private MockMvc mock;
     private static Faker faker;
@@ -36,19 +34,19 @@ public class ChangePaswordTest {
     private static String userPassword;
 
 
-
     @SuppressWarnings("MissortedModifiers")
     @BeforeClass
-    static public void init(){
+    static public void init() {
         faker = new Faker();
         requestBuilder = new TestRequestBuilder();
-        requestBuilder.init("oldPassword", "newPassword");
+        requestBuilder.init("userName", "userPassword", "userEmail");
     }
+
 
     public void createUser() throws Exception {
         this.mock.perform(
                 post("/api/user/signup")
-                        .contentType("application/json")
+                        .contentType("application/json")//MediaType.APPLICATION_JSON_VALUE)
                         .content(TestRequestBuilder.getJsonRequestForSignUp(userName, userPassword, userEmail)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
                 .andExpect(status().isOk())
@@ -58,81 +56,68 @@ public class ChangePaswordTest {
 
     @SuppressWarnings("ThrowInsideCatchBlockWhichIgnoresCaughtException")
     @Before
-    public void setUp(){
+    public void setUp() {
         userName = faker.name().username();
         userEmail = faker.internet().emailAddress();
         userPassword = faker.internet().password();
-        pathUrl = "/api/user/change_password";
+        pathUrl = "/api/user/signup";
 
         try {
             createUser();
-        }
-        catch (Exception ex) {
+        } catch (Exception ex) {
             throw new RuntimeException();
         }
     }
 
     @Test
-    public void changePasswordNormal() throws Exception{
+    public void signUpNormal() throws Exception {
         this.mock.perform(
                 post(pathUrl)
                         .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userPassword, faker.internet().password()))
-                        .sessionAttr("userName", userName))
+                        .content(requestBuilder.getJsonRequest(faker.name().username(),
+                                faker.internet().password(),
+                                faker.internet().emailAddress())))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.result", is(true)))
-                .andExpect(jsonPath("$.responseMessage", is("Ok! en")));
+                .andExpect(jsonPath("$.responseMessage", is("User created successfully! en")));
     }
 
-    @SuppressWarnings("InstanceMethodNamingConvention")
+
     @Test
-    public void changePasswordIncorrectOldPassword() throws Exception {
+    public void signUpConflictUserName() throws Exception {
         this.mock.perform(
                 post(pathUrl)
                         .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest("aaa", faker.internet().password()))
-                        .sessionAttr("userName", userName))
+                        .content(requestBuilder.getJsonRequest(userName, userPassword, userEmail)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isForbidden())
+                .andExpect(status().isConflict())
                 .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Invalid authentication data! en")));
+                .andExpect(jsonPath("$.responseMessage", is("Login already occupied! en")));
     }
 
-
     @Test
-    public void changePasswordNullOldPassword() throws Exception{
+    public void signUpConflictUserEmail() throws Exception {
+        final String otherUserName = faker.name().username();
+        final String otherUserPassword = faker.internet().password();
+
         this.mock.perform(
                 post(pathUrl)
                         .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(null, faker.internet().password()))
-                        .sessionAttr("userName", userName))
+                        .content(requestBuilder.getJsonRequest(otherUserName, otherUserPassword, userEmail)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isBadRequest())
+                .andExpect(status().isConflict())
                 .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Json contains null fields! en")));
+                .andExpect(jsonPath("$.responseMessage", is("Email already registered! en")));
     }
 
+
     @Test
-    public void changePasswordNullNewPassword() throws Exception{
+    public void signUpNullUserName() throws Exception {
         this.mock.perform(
                 post(pathUrl)
                         .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userPassword, null))
-                        .sessionAttr("userName", userName))
-                .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Json contains null fields! en")));
-    }
-
-    @Test
-    public void changePasswordNullBothPasswords() throws Exception{
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(null, null))
-                        .sessionAttr("userName", userName))
+                        .content(requestBuilder.getJsonRequest(null, userPassword, userEmail)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
                 .andExpect(status().isBadRequest())
                 .andExpect(jsonPath("$.result", is(false)))
@@ -140,33 +125,33 @@ public class ChangePaswordTest {
     }
 
     @Test
-    public void changePasswordNullSession() throws Exception{
+    public void signUpNullUserPassword() throws Exception {
         this.mock.perform(
                 post(pathUrl)
                         .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userPassword, faker.internet().password())))
+                        .content(requestBuilder.getJsonRequest(userName, null, userEmail)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isNotFound())
+                .andExpect(status().isBadRequest())
                 .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Invalid session! en")));
+                .andExpect(jsonPath("$.responseMessage", is("Json contains null fields! en")));
     }
 
+
     @Test
-    public void changePasswordInvalidSession() throws Exception {
+    public void signUpNullUserEmail() throws Exception {
         this.mock.perform(
                 post(pathUrl)
                         .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userPassword, faker.internet().password()))
-                        .sessionAttr("userName", faker.name().username()))
+                        .content(requestBuilder.getJsonRequest(userName, userPassword, null)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isNotFound())
+                .andExpect(status().isBadRequest())
                 .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Invalid session! en")));
+                .andExpect(jsonPath("$.responseMessage", is("Json contains null fields! en")));
     }
 
-    @SuppressWarnings("InstanceMethodNamingConvention")
+
     @Test
-    public void changePasswordIncorrectDocumentType() throws Exception{
+    public void signUpIncorrectDocumentType() throws Exception {
         this.mock.perform(
                 post(pathUrl)
                         .contentType("text/html"))
