+++ src/test/java/tests/Test/UserDataTest_IT.java
@@ -15,6 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.hamcrest.core.Is.is;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
@@ -23,29 +24,27 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest(classes = Application.class)
 @AutoConfigureMockMvc(print = MockMvcPrint.NONE)
 @Category(tests.IntegrationTest.class)
-public class SignInTest {
+//
+
+public class UserDataTest_IT {
     @Autowired
     private MockMvc mock;
-    private static TestRequestBuilder requestBuilder;
     private static Faker faker;
     private static String pathUrl;
     private static String userName;
     private static String userEmail;
     private static String userPassword;
 
-
     @SuppressWarnings("MissortedModifiers")
     @BeforeClass
     static public void init() {
         faker = new Faker();
-        requestBuilder = new TestRequestBuilder();
-        requestBuilder.init("userName", "userPassword");
     }
 
     public void createUser() throws Exception {
         this.mock.perform(
                 post("/api/user/signup")
-                        .contentType("application/json")//MediaType.APPLICATION_JSON_VALUE)
+                        .contentType("application/json")
                         .content(TestRequestBuilder.getJsonRequestForSignUp(userName, userPassword, userEmail)))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
                 .andExpect(status().isOk())
@@ -59,7 +58,7 @@ public class SignInTest {
         userName = faker.name().username();
         userEmail = faker.internet().emailAddress();
         userPassword = faker.internet().password();
-        pathUrl = "/api/user/signin";
+        pathUrl = "/api/user/data";
 
         try {
             createUser();
@@ -68,13 +67,11 @@ public class SignInTest {
         }
     }
 
-
     @Test
-    public void signInNormal() throws Exception {
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userName, userPassword)))
+    public void getDataOk() throws Exception {
+        mock.perform(
+                get(pathUrl)
+                        .sessionAttr("userName", userName))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.result", is(true)))
@@ -82,61 +79,24 @@ public class SignInTest {
     }
 
     @Test
-    public void signInNullUserName() throws Exception {
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(null, userPassword)))
-                .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Json contains null fields! en")));
-    }
-
-    @Test
-    public void signInNullUserPassword() throws Exception {
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userName, null)))
-                .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Json contains null fields! en")));
-    }
-
-
-    @Test
-    public void signInIncorrectUserName() throws Exception {
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest("Petya", userPassword)))
+    public void invalidSession() throws Exception {
+        mock.perform(
+                get(pathUrl)
+                        .sessionAttr("userName", faker.name().username()))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
                 .andExpect(status().isForbidden())
                 .andExpect(jsonPath("$.result", is(false)))
                 .andExpect(jsonPath("$.responseMessage", is("Invalid authentication data! en")));
     }
 
-
     @Test
-    public void signInIncorrectUserPassword() throws Exception {
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("application/json")
-                        .content(requestBuilder.getJsonRequest(userName, "no")))
+    public void nullSession() throws Exception {
+        mock.perform(
+                get(pathUrl))
                 .andExpect(content().contentType("application/json;charset=UTF-8"))
-                .andExpect(status().isForbidden())
+                .andExpect(status().isNotFound())
                 .andExpect(jsonPath("$.result", is(false)))
-                .andExpect(jsonPath("$.responseMessage", is("Invalid authentication data! en")));
+                .andExpect(jsonPath("$.responseMessage", is("Invalid session! en")));
     }
 
-
-    @Test
-    public void signInIncorrectDocumentType() throws Exception {
-        this.mock.perform(
-                post(pathUrl)
-                        .contentType("text/html"))
-                .andExpect(status().isUnsupportedMediaType());
-    }
 }
